+++ packages/svelte/tests/cases/asynclink/forward-props/index.js
@@ -1,6 +1,5 @@
 import { inMemory } from "@hickory/in-memory";
 import { createRouter, prepareRoutes } from "@curi/router";
-import { curiStores } from "@curi/svelte";
 
 import app from "./app.svelte";
 
@@ -13,11 +12,10 @@ const routes = prepareRoutes({
 });
 
 const router = createRouter(inMemory, routes);
-const stores = curiStores(router);
 
 export default function render() {
   const target = document.createElement("div");
-  new app.default({ target, props: { stores } });
+  new app.default({ target, props: { router } });
 
   const a = target.querySelector("a");
   expect(a.classList.contains("test")).toBe(true);