+++ packages/svelte/tests/cases/asynclink/query-hash/app.svelte
@@ -1,4 +1,4 @@
-<Router stores={stores}>
+<Router {router}>
   <AsyncLink name="Home" query="one=two" hash="test">Home</AsyncLink>
 </Router>
 
@@ -6,5 +6,5 @@
   import AsyncLink from "@curi/svelte/components/AsyncLink.svelte";
   import Router from "@curi/svelte/components/Router.svelte";
 
-  export let stores;
+  export let router;
 </script>