+++ packages/svelte/tests/cases/context/router/getter.svelte
@@ -1,0 +1,9 @@
+<script>
+  import { getRouter } from "@curi/svelte";
+
+  export let check;
+
+  const router = getRouter();
+
+  check(router);
+</script>