+++ packages/svelte/tests/cases/link/click/app.svelte
@@ -1,4 +1,4 @@
-<Router stores={stores}>
+<Router {router}>
   <Link name="User" params={{ id: "1" }}>User 1</Link>
 </Router>
 
@@ -6,5 +6,5 @@
   import Link from "@curi/svelte/components/Link.svelte";
   import Router from "@curi/svelte/components/Router.svelte";
 
-  export let stores;
+  export let router;
 </script>