+++ packages/svelte/tests/cases/link/query-hash/app.svelte
@@ -1,4 +1,4 @@
-<Router stores={stores}>
+<Router {router}>
   <Link name="Home" query="one=two" hash="test">Home</Link>
 </Router>
 
@@ -6,5 +6,5 @@
   import Link from "@curi/svelte/components/Link.svelte";
   import Router from "@curi/svelte/components/Router.svelte";
 
-  export let stores;
+  export let router;
 </script>