+++ packages/svelte/tests/cases/link/wrapper-destroy-callbacks/app.svelte
@@ -7,5 +7,5 @@
   import Router from "@curi/svelte/components/Router.svelte";
   import Wrapper from "./wrapper.svelte";
 
-  export let stores;
+  export let router;
 </script>