+++ packages/svelte/tests/cases/link/wrapper/app.svelte
@@ -1,4 +1,4 @@
-<Router stores={stores}>
+<Router {router}>
   <Link name="Home" wrapper={Wrapper}>Home</Link>
 </Router>
 
@@ -7,5 +7,5 @@
   import Router from "@curi/svelte/components/Router.svelte";
   import Wrapper from "./wrapper.svelte";
 
-  export let stores;
+  export let router;
 </script>